Extract `BaseClassProxyGenerator` base class
`ClassProxyGenerator` and `ClassProxyWithTargetGenerator` are now mostly
identical; extract the common bits into a new abstract base class.

This makes it easier to spot the actual differences between the two.

More importantly, it fixes the semantics of these types' hierarchy:
`ClassProxyGenerator` is a "without target" generator. It is wrong for
a "with target" generator to inherit from such a one as this contradicts
the base class' semantics.
