+++ src/Castle.Core/DynamicProxy/Generators/ClassProxyGenerator.cs
@@ -16,202 +16,34 @@ namespace Castle.DynamicProxy.Generators
 {
 	using System;
 	using System.Collections.Generic;
-	using System.Linq;
 	using System.Reflection;
 
 	using Castle.DynamicProxy.Contributors;
-	using Castle.DynamicProxy.Generators.Emitters;
 	using Castle.DynamicProxy.Generators.Emitters.SimpleAST;
-	using Castle.DynamicProxy.Internal;
 	using Castle.DynamicProxy.Serialization;
 
-	internal class ClassProxyGenerator : BaseProxyGenerator
+	internal sealed class ClassProxyGenerator : BaseClassProxyGenerator
 	{
 		public ClassProxyGenerator(ModuleScope scope, Type targetType, Type[] interfaces, ProxyGenerationOptions options)
 			: base(scope, targetType, interfaces, options)
 		{
-			EnsureDoesNotImplementIProxyTargetAccessor(targetType, "targetType");
 		}
 
-		private FieldReference TargetField => null;
+		protected override FieldReference TargetField => null;
 
 		protected override CacheKey GetCacheKey()
 		{
 			return new CacheKey(targetType, interfaces, ProxyGenerationOptions);
 		}
 
-		protected override Type GenerateType(string name, INamingScope namingScope)
-		{
-			IEnumerable<ITypeContributor> contributors;
-			var allInterfaces = GetTypeImplementerMapping(out contributors, namingScope);
-
-			var model = new MetaType();
-			// Collect methods
-			foreach (var contributor in contributors)
-			{
-				contributor.CollectElementsToProxy(ProxyGenerationOptions.Hook, model);
-			}
-			ProxyGenerationOptions.Hook.MethodsInspected();
-
-			var emitter = BuildClassEmitter(name, targetType, allInterfaces);
-
-			CreateFields(emitter);
-			CreateTypeAttributes(emitter);
-
-			// Constructor
-			var cctor = GenerateStaticConstructor(emitter);
-
-			var constructorArguments = new List<FieldReference>();
-
-			if (TargetField is { } targetField)
-			{
-				constructorArguments.Add(targetField);
-			}
-
-			foreach (var contributor in contributors)
-			{
-				contributor.Generate(emitter);
-
-				// TODO: redo it
-				if (contributor is MixinContributor mixinContributor)
-				{
-					constructorArguments.AddRange(mixinContributor.Fields);
-				}
-			}
-
-			// constructor arguments
-			var interceptorsField = emitter.GetField("__interceptors");
-			constructorArguments.Add(interceptorsField);
-			var selector = emitter.GetField("__selector");
-			if (selector != null)
-			{
-				constructorArguments.Add(selector);
-			}
-
-			GenerateConstructors(emitter, targetType, constructorArguments.ToArray());
-			GenerateParameterlessConstructor(emitter, targetType, interceptorsField);
-
-			// Complete type initializer code body
-			CompleteInitCacheMethod(cctor.CodeBuilder);
-
-			// Crosses fingers and build type
-
-			var proxyType = emitter.BuildType();
-			InitializeStaticFields(proxyType);
-			return proxyType;
-		}
-
-		private ClassProxyInstanceContributor GetProxyInstanceContributor(List<MethodInfo> methodsToSkip)
+		protected override ProxyInstanceContributor GetProxyInstanceContributor(List<MethodInfo> methodsToSkip)
 		{
 			return new ClassProxyInstanceContributor(targetType, methodsToSkip, interfaces, ProxyTypeConstants.Class);
 		}
 
-		private ClassProxyTargetContributor GetProxyTargetContributor(List<MethodInfo> methodsToSkip, INamingScope namingScope)
+		protected override CompositeTypeContributor GetProxyTargetContributor(List<MethodInfo> methodsToSkip, INamingScope namingScope)
 		{
 			return new ClassProxyTargetContributor(targetType, methodsToSkip, namingScope) { Logger = Logger };
-		}
-
-		protected virtual IEnumerable<Type> GetTypeImplementerMapping(out IEnumerable<ITypeContributor> contributors,
-																	  INamingScope namingScope)
-		{
-			var methodsToSkip = new List<MethodInfo>();
-			var proxyInstance = GetProxyInstanceContributor(methodsToSkip);
-			// TODO: the trick with methodsToSkip is not very nice...
-			var proxyTarget = GetProxyTargetContributor(methodsToSkip, namingScope);
-			IDictionary<Type, ITypeContributor> typeImplementerMapping = new Dictionary<Type, ITypeContributor>();
-
-			// Order of interface precedence:
-			// 1. first target
-			// target is not an interface so we do nothing
-
-			var targetInterfaces = targetType.GetAllInterfaces();
-			// 2. then mixins
-			var mixins = new MixinContributor(namingScope, false) { Logger = Logger };
-			if (ProxyGenerationOptions.HasMixins)
-			{
-				foreach (var mixinInterface in ProxyGenerationOptions.MixinData.MixinInterfaces)
-				{
-					if (targetInterfaces.Contains(mixinInterface))
-					{
-						// OK, so the target implements this interface. We now do one of two things:
-						if (interfaces.Contains(mixinInterface) && typeImplementerMapping.ContainsKey(mixinInterface) == false)
-						{
-							AddMappingNoCheck(mixinInterface, proxyTarget, typeImplementerMapping);
-							proxyTarget.AddInterfaceToProxy(mixinInterface);
-						}
-						// we do not intercept the interface
-						mixins.AddEmptyInterface(mixinInterface);
-					}
-					else
-					{
-						if (!typeImplementerMapping.ContainsKey(mixinInterface))
-						{
-							mixins.AddInterfaceToProxy(mixinInterface);
-							AddMappingNoCheck(mixinInterface, mixins, typeImplementerMapping);
-						}
-					}
-				}
-			}
-			var additionalInterfacesContributor = new InterfaceProxyWithoutTargetContributor(namingScope,
-																							 static (c, m) => NullExpression.Instance)
-			{ Logger = Logger };
-			// 3. then additional interfaces
-			foreach (var @interface in interfaces)
-			{
-				if (targetInterfaces.Contains(@interface))
-				{
-					if (typeImplementerMapping.ContainsKey(@interface))
-					{
-						continue;
-					}
-
-					// we intercept the interface, and forward calls to the target type
-					AddMappingNoCheck(@interface, proxyTarget, typeImplementerMapping);
-					proxyTarget.AddInterfaceToProxy(@interface);
-				}
-				else if (ProxyGenerationOptions.MixinData.ContainsMixin(@interface) == false)
-				{
-					additionalInterfacesContributor.AddInterfaceToProxy(@interface);
-					AddMapping(@interface, additionalInterfacesContributor, typeImplementerMapping);
-				}
-			}
-			// 4. plus special interfaces
-#if FEATURE_SERIALIZATION
-			if (targetType.IsSerializable)
-			{
-				AddMappingForISerializable(typeImplementerMapping, proxyInstance);
-			}
-#endif
-			try
-			{
-				AddMappingNoCheck(typeof(IProxyTargetAccessor), proxyInstance, typeImplementerMapping);
-			}
-			catch (ArgumentException)
-			{
-				HandleExplicitlyPassedProxyTargetAccessor(targetInterfaces);
-			}
-
-			contributors = new List<ITypeContributor>
-			{
-				proxyTarget,
-				mixins,
-				additionalInterfacesContributor,
-				proxyInstance
-			};
-			return typeImplementerMapping.Keys;
-		}
-
-		private void EnsureDoesNotImplementIProxyTargetAccessor(Type type, string name)
-		{
-			if (!typeof(IProxyTargetAccessor).IsAssignableFrom(type))
-			{
-				return;
-			}
-			var message =
-				string.Format(
-					"Target type for the proxy implements {0} which is a DynamicProxy infrastructure interface and you should never implement it yourself. Are you trying to proxy an existing proxy?",
-					typeof(IProxyTargetAccessor));
-			throw new ArgumentException(message, name);
 		}
 	}
 }
